requirements.txt
selenium>=4.15.2
pandas>=2.1.3
pyyaml>=6.0.1
webdriver-manager>=4.0.1